requirements.txt
--index-url https://download.pytorch.org/whl/cu118
torch
torchvision

ujson
pyyaml
matplotlib
torch_optimizer